Refactor group
